# internal/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Estoque microservice API documentation.",
        "title": "API Estoque",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/stock-items": {
            "get": {
                "description": "Pega todos os registros de item de estoque",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-items"
                ],
                "summary": "Listar items do estoque",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Faz a criação de item de estoque",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-items"
                ],
                "summary": "Cria item de estoque",
                "parameters": [
                    {
                        "description": "Stock Item",
                        "name": "stockItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stockitems.StockItems"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
        "/stock-items/{idWarehouse}/{idProduct}": {
            "get": {
                "description": "Retorna um item de estoque específico pelo idWarehouse e idProduct",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-items"
                ],
                "summary": "Buscar item de estoque por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do Warehouse",
                        "name": "idWarehouse",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID do Produto",
                        "name": "idProduct",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de um item de estoque existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-items"
                ],
                "summary": "Atualizar item de estoque",
                "parameters": [
                    {
                        "description": "Stock Item",
                        "name": "stockItem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/stockitems.StockItems"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Exclui um item de estoque pelo idWarehouse e idProduct",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stock-items"
                ],
                "summary": "Remover item de estoque",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID do Warehouse",
                        "name": "idWarehouse",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID do Produto",
                        "name": "idProduct",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpresponse.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httpresponse.Response": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "stockitems.StockItems": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "reserved": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
    }
}